move data related stuff to /data Fixes #9
